Create test file for Wait state handler
