+++ __tests__/StateMachine.test.ts
@@ -1,140 +1,12 @@
 import type { StateMachineDefinition } from '../src/typings/StateMachineDefinition';
 import { StateMachine } from '../src/StateMachine';
 import * as ChoiceHelperModule from '../src/ChoiceHelper';
-import * as utilModule from '../src/util';
 
 afterEach(() => {
   jest.clearAllMocks();
 });
 
 describe('State Machine', () => {
-  describe('Wait State', () => {
-    const mockSleepFunction = jest.fn();
-    const mockDateNowFunction = jest.fn(() => 1670198400000); // 2022-12-05T00:00:00Z
-
-    beforeEach(() => {
-      jest.spyOn(utilModule, 'sleep').mockImplementation(mockSleepFunction);
-      jest.spyOn(Date, 'now').mockImplementation(mockDateNowFunction);
-    });
-
-    test('should pause execution for the amount of seconds specified in the `Seconds` field', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'WaitState',
-        States: {
-          WaitState: {
-            Type: 'Wait',
-            Seconds: 10,
-            End: true,
-          },
-        },
-      };
-      const input = {};
-
-      const stateMachine = new StateMachine(definition);
-      await stateMachine.run(input);
-
-      expect(mockSleepFunction).toHaveBeenCalledWith(10000);
-    });
-
-    test('should pause execution until time in `Timestamp` field is reached', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'WaitState',
-        States: {
-          WaitState: {
-            Type: 'Wait',
-            Timestamp: '2022-12-05T05:45:00Z',
-            End: true,
-          },
-        },
-      };
-      const input = {};
-
-      const stateMachine = new StateMachine(definition);
-      await stateMachine.run(input);
-
-      expect(mockSleepFunction).toHaveBeenCalledWith(20700000);
-    });
-
-    test('should pause execution for the amount of seconds specified in the field referenced by `SecondsPath`', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            Parameters: { waitFor: 10 },
-            Next: 'WaitState',
-          },
-          WaitState: {
-            Type: 'Wait',
-            SecondsPath: '$.waitFor',
-            End: true,
-          },
-        },
-      };
-      const input = {};
-
-      const stateMachine = new StateMachine(definition);
-      await stateMachine.run(input);
-
-      expect(mockSleepFunction).toHaveBeenCalledWith(10000);
-    });
-
-    test('should pause execution until time specified in the field referenced by `TimestampPath` is reached', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            Parameters: { waitUntil: '2022-12-05T05:45:00Z' },
-            Next: 'WaitState',
-          },
-          WaitState: {
-            Type: 'Wait',
-            TimestampPath: '$.waitUntil',
-            End: true,
-          },
-        },
-      };
-      const input = {};
-
-      const stateMachine = new StateMachine(definition);
-      await stateMachine.run(input);
-
-      expect(mockSleepFunction).toHaveBeenCalledWith(20700000);
-    });
-
-    test('should pause execution for the specified amount of milliseconds if wait time override option is set', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'PassState',
-        States: {
-          PassState: {
-            Type: 'Pass',
-            Parameters: { waitUntil: '2022-12-05T05:45:00Z' },
-            Next: 'WaitState',
-          },
-          WaitState: {
-            Type: 'Wait',
-            TimestampPath: '$.waitUntil',
-            End: true,
-          },
-        },
-      };
-      const input = {};
-
-      const stateMachine = new StateMachine(definition);
-      await stateMachine.run(input, {
-        overrides: {
-          waitTimeOverrides: {
-            WaitState: 1500,
-          },
-        },
-      });
-
-      expect(mockSleepFunction).toHaveBeenCalledTimes(1);
-      expect(mockSleepFunction).toHaveBeenCalledWith(1500);
-    });
-  });
-
   describe('Choice State', () => {
     test('should transition to state that matches choice rule', async () => {
       const definition: StateMachineDefinition = {
